midilamp/requirements.txt
cython
pyserial
pyliblo
